requirements.txt
cbor2==5.6.4
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.7
luma.core==2.4.2
luma.emulator==1.5.0
luma.oled==3.13.0
pillow==10.3.0
pyftdi==0.55.4
pygame==2.5.2
pyserial==3.5
pyusb==1.2.1
requests==2.32.3
smbus2==0.4.3
typed_ast==1.5.5
urllib3==2.2.1
